+++ packages/topofit/CHANGELOG.md
@@ -5,6 +5,12 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @neurodesk/webapp-components@0.2.1
+
+
+### Patch Changes
+
+- Updated dependencies
   - @neurodesk/webapp-components@0.2.0
 
 ## 0.6.20260914